Add support for writing image direction tags.
When the GPX file includes a course (defined in the GPX 1.0
specification, but not later ones), it's used to write the GPSTrack and
GPSImgDirection EXIF tags. This is controlled by three command-line
flags: --heading, which enables writing GPSTrack, --direction, which
enables writing GPSImgDirection and specifies the number of degrees of
difference between the heading and the camera direction, and
--max-heading, which gives the maximum change in degrees between track
points, beyond which no EXIF tag will be written.