Fix a few tests to run in more locales
Some tests were missing the explicit setting of the locale that they
need to perform the correct comparison with golden test files. At the
same time, modify only the specific locale category needed for the test,
so help spot locale-dependent changes accidentally added to the code.
This won't work if someone specifies an odd locale with LC_ALL (since
that overrides the more specific variant), but normal systems won't be
doing that.

Add a CI test using a different locale to test this.