Warn if the times in the GPX file are not strictly increasing
This is an assumption of the matching algorithm to find the best
available location, so notify the user if it is untrue.

This warning showed that one of the test files (derived from a real GPX
file) unknowingly exhibited this problem, which combined with the
originating issue shows that perhaps this is a more common issue than
assumed.

This change does not check that times between track segments are also
strictly increasing, so it's possible that some bad files could still go
undetected, resulting in suboptimal matches if interpolation between
track segments is enabled.

Only the German translation is updated.

Fixes #27