Mention drag-and-drop being unavailable in Flatpak
This is due to the requirement of use of a portal which is not
supported in Gtk3 and likely never will be
(flatpak/xdg-desktop-portal#99). Adding permission to access the user's
home directory gets around this but opens a big hole in the sandbox.
It might still be worthwhile doing by default because of the added
functionality.